date_deleted producto

+++ public/app/js/project/product.datatable.js
@@ -1,7 +1,8 @@
 class ProductDatatable {
     ready() {
         var routeEdit = '/app/product/form';
-        var routeDelete = 'app/product/delete'
+        var routeDelete = 'app/product/delete';
+        var routeRestore = 'app/product/restore';
         var routeDatatable = '/app/product/datatable';
 
         var datatable = new CustomDatatable('#mt-product', {
@@ -44,6 +45,7 @@ class ProductDatatable {
                 { data: 'date_created', },
                 // Incluimos campos invisibles, útiles para filtros   
                 { data: 'market_ids', 'visible': false },
+                { data: 'product_status', 'visible': false },
                 {
                     data: null,
                     orderable: false,
@@ -51,11 +53,15 @@ class ProductDatatable {
                     className: 'text-center p-0',
                     width: 150,
                     render: function (data, type, full, meta) {
+                        if(full.product_status == 'Z') {
+                            return '<a class="btn btn-icon btn-active-light btn-sm p-3 profile-allow profile-A" href="javascript:AdminUtils.confirmRestore(\'' + routeRestore + '\', ' + data.id + ')" title="' + __('app.js.common.restore') + '"><i class="fas fa-undo fs-1 pb-1"></i></a>';
+                        }
+
                         var mRouteEdit = routeEdit + '/' + data.id;
                         var btnEdit = '<a class="btn btn-icon btn-active-light btn-sm p-3" href="' + mRouteEdit + '" title="' + __('app.js.common.edit') + '"><i class="fa-regular fa-pen-to-square fs-1 pb-1"></i></a>';
                         var btnDelete = '<a class="btn btn-icon btn-active-light btn-sm p-3 profile-allow profile-A" href="javascript:AdminUtils.confirmDelete(\'' + routeDelete + '\', ' + data.id + ')" title="' + __('app.js.common.delete') + '"><i class="fa-regular fa-trash-can fs-1 pb-1"></i></a>';
 
-                        if (full.total_booklets != 0 || full.total_references != 0 || full.id == EMPTY_PRODUCT) {
+                        if (full.id == EMPTY_PRODUCT) {
                             btnDelete = '';
                         }
